functions/requirements.txt
numpy
Pillow
deepFace
functions-framework
pgvector